Refactor Progress Bar into separate package
